correcting empty line bug by modifiing strmdup

+++ get_next_line.h
@@ -6,7 +6,7 @@
 /*   By: tkleynts <tkleynts@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/23 16:48:32 by tkleynts          #+#    #+#             */
-/*   Updated: 2019/11/12 18:28:51 by tkleynts         ###   ########.fr       */
+/*   Updated: 2019/11/14 10:05:20 by tkleynts         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@
 int			get_next_line(int fd, char **line);
 char		*endl_check(char *str);
 size_t		ft_strlen(const char *str);
-char		*ft_strmdup(char *s1, size_t size);
+char		*ft_strmdup(char *s1, int size);
 char		*ft_strjoin(char *line, char *buff);
 
 #endif
